pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "t-tool"
version = "0.1.0"
description = "Local-first translation tooling foundation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "sqlmodel>=0.0.16",
  "typer>=0.12.3",
  "pydantic>=2.7",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
]

[project.scripts]
tt = "tt_cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["tt_core", "tt_cli"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
